requirements.txt
pandas>=1.5.0
fpdf2>=2.7.0
tqdm>=4.64.0
openpyxl>=3.0.0
tkinter
